requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
attrs==25.4.0
click==8.3.0
colorama==0.4.6
fastapi==0.121.1
frozenlist==1.8.0
h11==0.16.0
idna==3.11
magic-filter==1.0.12
maxapi==0.9.7
multidict==6.7.0
propcache==0.4.1
psycopg2-binary==2.9.11
puremagic==1.30
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
pytz==2025.1
uvicorn==0.38.0
yarl==1.22.0
